Move cleanup-index command to fixes
