+++ hepdata/ext/elasticsearch/api.py
@@ -27,7 +27,7 @@ from elasticsearch.exceptions import TransportError
 from elasticsearch_dsl import Search
 from elasticsearch_dsl.query import QueryString, Q
 from invenio_pidstore.models import RecordIdentifier
-from sqlalchemy import func, and_, distinct
+from sqlalchemy import and_
 from sqlalchemy.orm import aliased
 
 
@@ -259,64 +259,6 @@ def reindex_batch(hepsubmission_record_ids, index):
 
     log.info('Finished indexing, now pushing data keywords\n######')
     push_data_keywords(pub_ids=indexed_publications)
-
-
-@default_index
-def cleanup_index_all(index=None, batch=5, synchronous=False):
-    # Find entries in elasticsearch which are from previous versions of submissions and remove
-    # Get all finished HEPSubmission ids with version numbers less than the max
-    # finished version by doing a left outer join of hepsubmission with itself
-    h1 = aliased(HEPSubmission)
-    h2 = aliased(HEPSubmission)
-
-    qry = db.session.query(distinct(h1.id)) \
-        .join(h2,
-              and_(h1.publication_recid == h2.publication_recid,
-                   h1.version < h2.version,
-                   h2.overall_status == 'finished'),
-              isouter=True) \
-        .filter(h2.id != None) \
-        .order_by(h1.id)
-    res = qry.all()
-    ids = [x[0] for x in res]
-
-    count = 0
-    while count < len(ids):
-        batch_ids = ids[count:min(count + batch, len(ids))]
-        if synchronous:
-            cleanup_index_batch(batch_ids, index)
-        else:
-            print('Sending batch of IDs {0} to {1} to celery'.format(batch_ids[0], batch_ids[-1]))
-            cleanup_index_batch.delay(batch_ids, index)
-        count += batch
-
-
-@shared_task
-def cleanup_index_batch(hepsubmission_record_ids, index):
-    log.info('Cleaning up index for data records for hepsubmission IDs {0} to {1}'.format(hepsubmission_record_ids[0], hepsubmission_record_ids[-1]))
-    # Find all datasubmission entries matching the given hepsubmission ids,
-    # where the version is not the highest version present (i.e. there is not
-    # a v2 record with the same associated_recid)
-    d1 = aliased(DataSubmission)
-    d2 = aliased(DataSubmission)
-    qry = db.session.query(d1.associated_recid) \
-        .join(HEPSubmission,
-              and_(HEPSubmission.publication_recid == d1.publication_recid,
-                   HEPSubmission.version == d1.version),
-              isouter=True) \
-        .join(d2,
-              and_(d1.associated_recid == d2.associated_recid,
-                   d1.version < d2.version),
-              isouter=True) \
-        .filter(HEPSubmission.id.in_(hepsubmission_record_ids), d2.id == None) \
-        .order_by(d1.id)
-    res = qry.all()
-
-    ids = [x[0] for x in res]
-    if ids:
-        log.info(f'Deleting entries from index with ids {ids}')
-        s = RecordsSearch(index=index).filter('terms', _id=ids)
-        s.delete()
 
 
 @default_index
